Refactors folder structure
