+++ ci/app/spoken/spoken-dialect-parser.ts
@@ -1,7 +1,7 @@
-import { SpokenDialectPhoneme } from "../../src/phonemes/spoken/spoken-phoneme";
-import { DataParser, DataWarning } from "./data-parser";
-import { getSeperatedValueData, getUniqueValues } from "./parse-utils";
-import { SpokenDialect } from "../../src/phonemes/spoken/spoken-dialect";
+import { SpokenDialectPhoneme } from "../../../src/phonemes/spoken/spoken-phoneme";
+import { DataParser, DataWarning } from "../data-parser";
+import { getSeperatedValueData, getUniqueValues } from "../parse-utils";
+import { SpokenDialect } from "../../../src/phonemes/spoken/spoken-dialect";
 
 type RawData = {
     language: string,
